update tag component
