+++ frontend/src/components/Tag.tsx
@@ -1,13 +1,15 @@
 interface TagValues {
-  tag: string[];
+  tag: string;
 }
 
 const Tag = ({ tag }: TagValues) => {
+  const tagsArray = tag.match(/#[^\s]+/g)?.map((t) => t.slice(1));
+
   return (
     <div className="flex gap-2 items-center px-2">
-      {tag.map((t, i) => (
+      {tagsArray?.map((tag, i) => (
         <div key={i} className="text text-xs tag-border p-1">
-          {t}
+          {tag}
         </div>
       ))}
     </div>
